Rename PropertyTest to ExUnitProperties and move "gen all" to it
