+++ lib/stream_data.ex
@@ -75,7 +75,7 @@ defmodule StreamData do
 
   Shrinking is used for property testing, where it's important to find the
   minimal value for which a property fails (since generated values are often
-  bigger and full of garbage). See `PropertyTest` for more information.
+  bigger and full of garbage). See `ExUnitProperties` for more information.
 
   Note that the generation size is not related in any way with shrinking: while
   intuitively one may think that shrinking just means decreasing the generation
@@ -1676,116 +1676,6 @@ defmodule StreamData do
           {:error, reason} ->
             shrink_failure(shrink_initial_cont(child.children), cont, reason, fun, nodes_visited + 1, config)
         end
-    end
-  end
-
-  @doc """
-  Syntactic sugar to create generators.
-
-  This macro provides ad hoc syntax to write complex generators. Let's see a
-  quick example to get a feel of how it works. Say we have a `User` struct:
-
-      defmodule User do
-        defstruct [:name, :email]
-      end
-
-  We can create a generator of users like this:
-
-      email_generator = map({binary(), binary()}, fn left, right -> left <> "@" <> right end)
-      gen all name <- binary(),
-              email <- email_generator do
-        %User{name: name, email: email}
-      end
-
-  Everything between `gen all` and `do` is referred to as **clauses**. Clauses
-  are used to specify the values to generate to be used in the body. The newly
-  created generator will generated values that are the return value of the
-  `do` body for the generated values from the clauses.
-
-  ### Clauses
-
-  As seen in the example above, clauses can be of three types:
-
-    * value generation - they have the form `pattern <- generator` where
-      `generator` must be a generator. These clauses take a value out of
-      `generator` on each run and match it against `pattern`. Variables bound in
-      `pattern` can be then used throughout subsequent clauses and in the `do`
-      body.
-
-    * binding - they have the form `pattern = expression`. They are exactly like
-      assignment through the `=` operator: if `pattern` doesn't match
-      `expression`, an error is raised. They can be used to bind values for use
-      in subsequent clauses and in the `do` body.
-
-    * filtering - they have the form `expression`. If a filtering clause returns
-      a truthy value, then the set of generated values that appear before the
-      filtering clause is considered valid and generation continues. If the
-      filtering clause returns a falsey value, then the current value is
-      considered invalid and a new value is generated. Note that filtering
-      clauses should not filter out too many times; in case they do, a
-      `StreamData.FilterTooNarrowError` error is raised (same as `filter/3`).
-
-  ### Body
-
-  The return value of the body passed in the `do` block is what is ultimately
-  generated by the generator return by this macro.
-
-  ## Shrinking
-
-  See the module documentation for more information on shrinking. Clauses affect
-  shrinking in the following way:
-
-    * binding clauses don't affect shrinking
-    * filtering clauses affect shrinking like `filter/3`
-    * value generation clauses affect shrinking similarly to `bind/2`
-
-  """
-  defmacro gen({:all, _meta, clauses} = _generation_clauses, [do: body] = _block) do
-    compile(clauses, body)
-  end
-
-  defp compile(clauses, body) do
-    quote do
-      var!(generated_values, unquote(__MODULE__)) = []
-      {:cont, data} = unquote(compile_clauses(clauses, body))
-      data
-    end
-  end
-
-  defp compile_clauses([], body) do
-    quote do
-      var!(generated_values, unquote(__MODULE__)) = Enum.reverse(var!(generated_values, unquote(__MODULE__)))
-      {:cont, StreamData.constant(unquote(body))}
-    end
-  end
-
-  defp compile_clauses([{:<-, _meta, [pattern, generator]} = clause | rest], body) do
-    quote do
-      data =
-        StreamData.bind_filter(unquote(generator), fn unquote(pattern) = generated_value ->
-          var!(generated_values, unquote(__MODULE__)) =
-            [{unquote(Macro.to_string(clause)), generated_value} | var!(generated_values, unquote(__MODULE__))]
-          unquote(compile_clauses(rest, body))
-        end)
-
-      {:cont, data}
-    end
-  end
-
-  defp compile_clauses([{:=, _meta, [_left, _right]} = assignment | rest], body) do
-    quote do
-      unquote(assignment)
-      unquote(compile_clauses(rest, body))
-    end
-  end
-
-  defp compile_clauses([clause | rest], body) do
-    quote do
-      if unquote(clause) do
-        unquote(compile_clauses(rest, body))
-      else
-        :skip
-      end
     end
   end
